move link from footer to header
