+++ packages/client/hmi-client/src/assets/css/theme/_variables.scss
@@ -946,6 +946,8 @@ $petri-inputBox: var(--surface-0);
 	--text-color-danger: #{$dangerTextButton};
 	--constrain-width: 55rem;
 	color-scheme: light;
+	--gap: 1rem;
+	--gap-small: .5rem;
 	--petri-nodeBorder: var(--gray-300);
 	--petri-nodeFill: var(--gray-300);
 	--petri-lineColor: #{$petri-lineColor};
