+++ packages/client/hmi-client/src/assets/css/theme/designer/components/misc/_avatar.scss
@@ -1,6 +1,5 @@
 .p-avatar {
     background-color: $avatarBg;
-    border-radius: $borderRadius;
 
     &.p-avatar-lg {
         width: 3rem;
